requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
PyPDF2
python-docx
